# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample serice for managing orders",
        "title": "Flow-Diet-Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "soberkoder@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/calendar/recipes/create": {
            "post": {
                "description": "User creates a recipe in the calendar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Create Recipes In User Calendar",
                "parameters": [
                    {
                        "description": "Create Recipes In Calendar",
                        "name": "calendar",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/calendar.Calendar"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/calendar/recipes/delete": {
            "delete": {
                "description": "Delete Recipe In User Calendar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Delete Recipe In User Calendar",
                "parameters": [
                    {
                        "description": "Delete Recipe In Calendar. Please Use thisfor send request: {'ID':1}",
                        "name": "calendar",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/calendar.Calendar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/calendar/recipes/update": {
            "patch": {
                "description": "Update Recipe In User Calendar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Update Recipe In User Calendar",
                "parameters": [
                    {
                        "description": "Update Recipe In Calendar",
                        "name": "calendar",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/calendar.Calendar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/materials/create": {
            "post": {
                "description": "Create A New Material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "Create a new material",
                "parameters": [
                    {
                        "description": "Create New Material",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.Material"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/materials/delete": {
            "delete": {
                "description": "Delete Material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "Delete Material",
                "parameters": [
                    {
                        "description": "Delete a Material",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.Material"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/materials/get": {
            "post": {
                "description": "Get a material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "Get a material",
                "parameters": [
                    {
                        "description": "Get Material",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.Material"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/materials/update": {
            "patch": {
                "description": "Update Material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "Update Material",
                "parameters": [
                    {
                        "description": "Update a Material",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.Material"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/recipes/create": {
            "post": {
                "description": "Create A New Recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Create a new recipe",
                "parameters": [
                    {
                        "description": "Create New Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recipe.Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/recipes/delete": {
            "delete": {
                "description": "Delete Recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Delete Recipe",
                "parameters": [
                    {
                        "description": "Delete a Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recipe.Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/recipes/get": {
            "post": {
                "description": "Get a recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Get a recipe",
                "parameters": [
                    {
                        "description": "Get Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recipe.Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/recipes/update": {
            "patch": {
                "description": "Update Recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Update Recipe",
                "parameters": [
                    {
                        "description": "Update a Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recipe.Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/recommendation/getrecipes": {
            "post": {
                "description": "Get Recommendations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendation"
                ],
                "summary": "Get Recommendations",
                "parameters": [
                    {
                        "description": "dates(epoch format) : {'start_date':1643914403 ,'end_date':1644173603 }",
                        "name": "recommendation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/search/recipes": {
            "post": {
                "description": "Search Recipes by Title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search Recipes",
                "parameters": [
                    {
                        "description": "Please write search word directly",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
        "calendar.Calendar": {
            "type": "object",
            "properties": {
                "date_epoch": {
                    "type": "integer",
                    "example": 1643743444
                },
                "meal": {
                    "type": "string",
                    "example": "breakfast"
                },
                "prepared": {
                    "type": "boolean",
                    "example": false
                },
                "recipe_id": {
                    "type": "integer",
                    "example": 1
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}